first step of refactoring
