+++ reimaging/upload.go
@@ -20,11 +20,11 @@ var title string
 func init() {
 	rootCmd.AddCommand(uploadCmd)
 
-	uploadCmd.Flags().IntVarP(&AlbumId, "album-id", "", 0, "Using an existing album to upload")
+	uploadCmd.Flags().IntVarP(&AlbumID, "album-id", "", 0, "Using an existing album to upload")
 	uploadCmd.Flags().StringVarP(&title, "title", "t", "", "Create new vk album with title and uploads")
 }
 
 func upload(args []string) {
 	fmt.Println("upload called with", args[0])
-	fmt.Println(AlbumId, title)
+	fmt.Println(AlbumID, title)
 }
